split up physmon workflows into separate files
