--- CI/physmon/workflows/truth_tracking_kalman.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python3
-
-from physmon_common import makeSetup
-
-setup = makeSetup()
-
-print(setup)
